requirements.txt
deep-translator
openai
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor
wget
pycountry
psutil
pytube
pykeyboard
unidecode
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
urllib3
flask
pyshorteners
youtube-search
youtube-search-python
async_pymongo
gpytranslate
search_engine_parser
emojis
cloudscraper
blackpink
pycryptodome
pickledb
pyfiglet
qrcode
pytz
pydub
apscheduler
gtts

httpx==0.27.2
gunicorn==22.0.0
pillow==11.0.0
pymongo==4.9.1
py-tgcalls==0.9.7
lexica-api==1.5.7
#pymongo==3.12.0

git+https://github.com/Mayuri-Chan/pyrofork
git+https://github.com/yt-dlp/yt-dlp
git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master